Removed unnecessary port_loadNewThread function.
In osThreadDelete(), after the process stack is freed, the stack will
still push some data onto its process stack. But this will not corrupt
the heap. Because the system is still in the critical section, the freed
block can only be splitted by another thread mallocing. The stack push
or context save will not corrput the block links in the heap.